Refactor sanity checks for cleaner layout
